requirements.txt
Django==3.0.6
pytube3==9.6.4
requests==2.23.0

